Wiring up Add Track Modal to search results
Add Track modal now appears when clicking on a track in search results.
Also moved all functions related to searching / adding tracks to an
addTrack.js file. And locked it down so this file is only included if
the user is logged in (unauthenticated users wont be able to search /
add tracks and will not be able to access /Scripts/addTrack.js file